Ajuste de bugs
